+++ account_mass_reconcile_by_purchase_line/README.rst
@@ -24,7 +24,7 @@ To use this module, you need to:
 
 .. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
    :alt: Try me on Runbot
-   :target: https://runbot.odoo-community.org/runbot/98/10.0
+   :target: https://runbot.odoo-community.org/runbot/98/11.0
 
 Bug Tracker
 ===========
