+++ account_mass_reconcile_by_purchase_line/__init__.py
@@ -1,6 +1,1 @@
-# -*- coding: utf-8 -*-
-# © 2015-17 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-
 from . import models
-from . import wizards
