+++ account_mass_reconcile_by_purchase_line/__manifest__.py
@@ -1,11 +1,10 @@
-# -*- coding: utf-8 -*-
-# © 2015-17 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
+# © 2015-18 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 {
     "name": "Account Mass Reconcile by Purchase Line",
     "summary": "Allows to reconcile based on the PO line",
-    "version": "10.0.1.0.0",
+    "version": "11.0.1.0.0",
     "author": "Eficent, "
               "Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/bank-statement-reconcile",
