+++ account_mass_reconcile_by_purchase_line/models/__init__.py
@@ -1,6 +1,3 @@
-# -*- coding: utf-8 -*-
-# © 2015-17 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-
 from . import mass_reconcile
 from . import base_advanced_reconciliation
+from . import simple_reconciliation
