+++ account_mass_reconcile_by_purchase_line/models/mass_reconcile.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# © 2015-17 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
+# © 2015-18 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
 from odoo import api, models
@@ -9,8 +8,9 @@ class AccountMassReconcileMethod(models.Model):
     _inherit = 'account.mass.reconcile.method'
 
     @api.model
-    def _get_all_rec_method(self):
-        methods = super(AccountMassReconcileMethod, self)._get_all_rec_method()
+    def _get_reconcilation_methods(self):
+        methods = super(AccountMassReconcileMethod,
+                        self)._get_reconcilation_methods()
         methods += [
             ('mass.reconcile.advanced.by.purchase.line',
              'Advanced. Product, purchase order line.'),
