--- account_mass_reconcile_by_purchase_line/wizards/__init__.py
@@ -1,5 +1,0 @@
-# -*- coding: utf-8 -*-
-# © 2015-17 Eficent Business and IT Consulting Services S.L. (www.eficent.com)
-# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
-
-from . import advanced_reconciliation
